search/index: fixed a number of bad conversions to properly deal with text documents bigger than 2GB

+++ src/internfile/mh_mbox.cpp
@@ -108,8 +108,8 @@ public:
             return -1;
         }
         if (fseeko(fp, cacheoffset(msgnum), SEEK_SET) != 0) {
-            LOGDEB0(("MboxCache::get_offsets: seek %ld errno %d\n",
-                     (long)cacheoffset(msgnum), errno));
+            LOGDEB0(("MboxCache::get_offsets: seek %lld errno %d\n",
+                     cacheoffset(msgnum), errno));
             return -1;
         }
         mbhoff_type offset = -1;
